test: check that statsd is unrefed
Node has no documented way to unref a child_process with an ipc channel,
so this depends on the undocumented `_channel` property of a
`ChildProcess` object.

Etsy seems to not like when it exits before being configured, it
attempts to flush but finds configuration missing, and errors out. Since
the child's stderr is discarded (actually, logged via `debug()`), this
shouldn't matter.